Move files to flatten structure and reduce layers
