--- apps/admin-web/src/components/pages/index.ts
@@ -1,2 +1,0 @@
-import { ToolList } from "./ToolList";
-export { ToolList };
