--- apps/admin-web/src/components/pages/tools/index.ts
@@ -1,10 +1,0 @@
-import { Events } from "./Events";
-/*
-import { Membership } from "./Membership";
-import { Products } from "./Products";
-import { Sigs } from "./Sigs";
-import { Sponsors } from "./Sponsors";
-export { Membership, Events, Sponsors, Sigs, Products };
- */
-
-export { Events };
